Gemini-Models-Object-detection/requirements.txt
google-genai
pillow
gdown
python-dotenv
pydantic
requests
matplotlib
